Refactored LED code
