+++ src/app.js
@@ -1,8 +1,18 @@
+const path = require( 'path' );
+
 const config = require( './config-manager' );
 let ledScripts = require( "./led-scripts/led-scripts.js" );
 const matrixScripts = require( "./led-scripts/matrix-scripts.js" );
 const processSubgroups = require( "./subgroups.js" );
-const path = require( 'path' );
+const {
+    blankStrip,
+    clearStrip,
+    disableVirtualStripSegments,
+    markControllersUpdated,
+    resetVirtualStrips,
+    stripIndexTextKey,
+    writeConfigToStrips
+} = require( "./ledControl" );
 
 //grab data from config
 let features = config.get( "features" );
@@ -23,7 +33,6 @@ let weatherData = {
     "outdoor": {}
 };
 let currentLEDs = [];
-let virtualStrips = [];
 let controllers = [];
 let controllerUpdates = [];
 let drawOnInterval = false;
@@ -34,6 +43,7 @@ let enableLiveView = false;
 let liveViewTimeout = null;
 let emitLiveViewUpdate = null;
 
+// Helper Function
 function setLiveViewEnabled() {
     enableLiveView = true;
     if ( liveViewTimeout ) {
@@ -48,128 +58,7 @@ function setLiveViewEmitter( emitter ) {
     emitLiveViewUpdate = emitter;
 }
 
-function markControllersUpdated( stripConfiguration, stripIndex ) {
-    if ( stripConfiguration?.controllers?.length ) {
-        stripConfiguration.controllers.forEach( controller => {
-            controllerUpdates[controller] = true;
-        } );
-        return;
-    }
-    if ( stripConfiguration?.controller !== undefined ) {
-        controllerUpdates[stripConfiguration.controller] = true;
-        return;
-    }
-    if ( Array.isArray( stripIndex ) ) {
-        controllerUpdates[stripIndex[0]] = true;
-    }
-}
-
-function clearStripEffects( strip ) {
-    if ( strip.effect ) {
-        clearInterval( strip.effectTimout );
-        strip.effect = {};
-        strip.effectTimout = null;
-    }
-    if ( strip.transition ) {
-        clearInterval( strip.transition.interval );
-        strip.transition = null;
-    }
-}
-
-function resetVirtualStrips() {
-    virtualStrips.forEach( vStrip => clearStripEffects( vStrip ) );
-    virtualStrips = [];
-}
-
-function stripIndexTextKey( stripId ) {
-    return `${ stripId[0] }.${ stripId[1] }`;
-}
-
-function resetVirtualStripSegments( vStrip ) {
-    vStrip.segments.forEach( segment => {
-        segment.enabled = true;
-    } );
-}
-
-function buildVirtualStrip( groupIndex, strips ) {
-    const segments = [];
-    let totalLength = 0;
-    const controllersSet = new Set();
-    ( strips || [] ).forEach( ( stripId ) => {
-        const controllerIndex = stripId[0];
-        const stripIndex = stripId[1];
-        const stripConfiguration = controllersConfig?.[controllerIndex]?.strips?.[stripIndex];
-        if ( !stripConfiguration ) {
-            return;
-        }
-        const length = stripConfiguration.configuredLength ?? stripConfiguration.length ?? 0;
-        segments.push( {
-            id: stripId,
-            controller: controllerIndex,
-            stripIndex,
-            offset: totalLength,
-            length,
-            enabled: true
-        } );
-        totalLength += length;
-        controllersSet.add( controllerIndex );
-        const target = currentLEDs?.[controllerIndex]?.[stripIndex];
-        if ( target ) {
-            clearStripEffects( target );
-        }
-    } );
-
-    const vStrip = blankStrip( {
-        id: [ "sharedRender", groupIndex ],
-        name: `Shared ${ groupIndex }`,
-        length: totalLength,
-        controller: null
-    } );
-
-    vStrip.type = "sharedRender";
-    vStrip.group = groupIndex;
-    vStrip.strips = strips;
-    vStrip.segments = segments;
-    vStrip.controllers = Array.from( controllersSet );
-    vStrip.applyToStrips = ( arr ) => {
-        vStrip.segments.forEach( segment => {
-            if ( !segment.enabled ) {
-                return;
-            }
-            const colors = arr.slice( segment.offset, segment.offset + segment.length );
-            const target = currentLEDs?.[segment.controller]?.[segment.stripIndex];
-            if ( target ) {
-                target.arr = colors;
-                target.trigger = vStrip.trigger || target.trigger;
-            }
-        } );
-    };
-
-    return vStrip;
-}
-
-function disableVirtualStripSegments( stripId, keepGroup = null ) {
-    const key = stripIndexTextKey( stripId );
-    virtualStrips.forEach( ( vStrip ) => {
-        if ( keepGroup !== null && vStrip.group === keepGroup ) {
-            return;
-        }
-        let enabledCount = 0;
-        vStrip.segments.forEach( segment => {
-            if ( stripIndexTextKey( segment.id ) === key ) {
-                segment.enabled = false;
-            }
-            if ( segment.enabled ) {
-                enabledCount += 1;
-            }
-        } );
-        if ( enabledCount === 0 ) {
-            clearStripEffects( vStrip );
-        }
-    } );
-    virtualStrips = virtualStrips.filter( vStrip => vStrip.segments.some( segment => segment.enabled ) );
-}
-
+// Setup LED internals
 function rebuildControllers( controllersList, options = {} ) {
     const { exitOnFailure = false } = options;
     resetVirtualStrips();
@@ -350,7 +239,7 @@ if ( features.hostWebControl || features.webAPIs || features.gpioButtonsOnWeb ) 
             clearAppConfigs,
             setStripDefaults,
             drawLEDs,
-            writeConfigToStrips,
+            writeConfigToStrips: writeConfigToStripsWithContext,
             reloadLEDScripts,
             rebuildControllers,
             setHomekitConfig: setupHomekit,
@@ -417,6 +306,7 @@ function destroyHomekitInstances() {
     homekitInstances = [];
 }
 
+// Set up homekit
 function setupHomekit( homeKitConfig ) {
     const nextConfig = homeKitConfig || [];
     if ( !features.homekit ) {
@@ -434,38 +324,7 @@ function setupHomekit( homeKitConfig ) {
     return rebuiltConfig;
 }
 
-//set up homekit
 setupHomekit( config.get( "homekit" ) );
-
-//set up matrix
-let matrixInterval;
-
-if ( features.matrixDisplay ) {
-    setTimeout( () => {
-        changeMatrix( { 'id': displayMatrix.default } )
-    }, 500 );
-}
-
-function newLEDarr( size, color ) {
-    let arr = [];
-    for ( let i = 0; i < size; i++ ) {
-        arr.push( color );
-    }
-    return arr;
-}
-
-function clearAppConfigs( noClear ) {
-    noClear = noClear.map( s => stripIndexTextKey(s) ) ?? [];
-    currentLEDs.forEach( ( controller, cIndex ) => {
-        controller.forEach( ( strip, sIndex ) => {
-            if ( strip.trigger === "app" && !noClear.includes( stripIndexTextKey([cIndex, sIndex]) ) ) {
-                clearStrip( strip );
-                disableVirtualStripSegments( [ cIndex, sIndex ] );
-                controllerUpdates[cIndex] = true;
-            }
-        } );
-    } );
-}
 
 function reloadLEDScripts() {
     console.log( "reloading LED scripts" );
@@ -475,175 +334,19 @@ function reloadLEDScripts() {
     ledScripts = require( "./led-scripts/led-scripts" );
 }
 
-function blankStrip( strip ) {
-    return {
-        "id": strip.id,
-        "name": strip.name,
-        "length": strip.length,
-        "controller": strip.controller,
-        "arr": newLEDarr( strip.length, "000000" ),
-        "trigger": "",
-        "effect": {},
-        "effectTimout": null,
-        "default": strip.default
-    };
-}
-
-function clearStrip( currentStrip, options = {} ) {
-    let oldColorArr = [];
-    clearStripEffects( currentStrip );
-    if ( options.transition ) {
-        oldColorArr = currentStrip.arr;
-        if ( oldColorArr.length === 0 ) {
-            oldColorArr = newLEDarr( currentStrip.length, "000000" );
-        }
-    }
-
-    // Clear strip config completely
-    Object.assign( currentStrip, blankStrip( currentStrip ) );
-
-    return oldColorArr;
-}
-
-function writeConfigToStrips( stripIndex, options ) {
-    let currentStrip, oldColorArr, stripConfiguration;
-
-    // Clear any outstanding effects and prep strip
-    if ( stripIndex.type === "sharedRender" ) {
-        // We're dealing with a virtual strip, we need to set one up
-
-        // Check to see if any virtual strip is already going with the same group (stripIndex.group is equal)
-        // If it is, clear and use that strip
-        const existingVStrip = virtualStrips.find( vs => vs.group === stripIndex.group );
-        if ( existingVStrip ) {
-            const existingKeys = ( existingVStrip.strips || [] ).map( stripIndexTextKey ).join( "|" );
-            const nextKeys = ( stripIndex.strips || [] ).map( stripIndexTextKey ).join( "|" );
-            if ( existingKeys !== nextKeys ) {
-                clearStripEffects( existingVStrip );
-                virtualStrips = virtualStrips.filter( vStrip => vStrip !== existingVStrip );
-                currentStrip = buildVirtualStrip( stripIndex.group, stripIndex.strips );
-                virtualStrips.push( currentStrip );
-            } else {
-                oldColorArr = clearStrip( existingVStrip, options );
-                resetVirtualStripSegments( existingVStrip );
-                currentStrip = existingVStrip;
-            }
-        } else {
-            // If there isn't an existing virtual strip, create one
-            currentStrip = buildVirtualStrip( stripIndex.group, stripIndex.strips );
-            virtualStrips.push( currentStrip );
-        }
-
-        currentStrip.segments?.forEach( segment => {
-            const target = currentLEDs?.[segment.controller]?.[segment.stripIndex];
-            if ( target ) {
-                clearStripEffects( target );
-            }
-        } );
-        currentStrip.strips.forEach( stripId => disableVirtualStripSegments( stripId, stripIndex.group ) );
-
-        stripConfiguration = {
-            length: currentStrip.length,
-            configuredLength: currentStrip.length,
-            controllers: currentStrip.controllers
-        };
-    } else {
-        // Normal strip, set and clear
-        currentStrip = currentLEDs[stripIndex[0]][stripIndex[1]];
-        oldColorArr = clearStrip( currentStrip, options );
-        stripConfiguration = controllersConfig[stripIndex[0]].strips[stripIndex[1]];
-
-        // Check to see if this strip was a part of any virtual strips
-        // If it is, disable just the strip on the vStrip
-        // If there are no enabled parts of the vStrip, clear any intervals and delete it
-        disableVirtualStripSegments( stripIndex );
-    }
-
-    // Generate pattern
-    currentStrip.arr = ledScripts.patterns[options.pattern].generate( stripConfiguration.configuredLength ?? stripConfiguration.length, options.patternOptions );
-    //set trigger
-    currentStrip.trigger = options.trigger;
-    if ( options.transition ) {
-        let newColorArr = currentStrip.arr;
-        currentStrip.arr = oldColorArr;
-        currentStrip.transition = new ledScripts.transitions[options.transition].Create( newColorArr, oldColorArr, options.transitionOptions, MAX_FPS );
-        currentStrip.transition.interval = setInterval( () => {
-            if ( currentStrip.transition ) {
-                currentStrip.transition.step( ( arr ) => {
-                    currentStrip.arr = arr;
-                    markControllersUpdated( stripConfiguration, stripIndex );
-                    if ( currentStrip.type === "sharedRender" ) {
-                        currentStrip.applyToStrips( currentStrip.arr );
-                    }
-                    drawLEDs();
-                } )
-            }
-        }, currentStrip.transition.intervalTime );
-    }
-    //if there is an effect, apply & set it up
-    if ( options.effect ) {
-        //create the new effect
-        currentStrip.effect = new ledScripts.effects[options.effect].Create(
-            currentStrip.arr,
-            {
-                ...options.effectOptions,
-                numLEDs: stripConfiguration.configuredLength ?? stripConfiguration.length
-            } );
-
-        // Run the effect once - Needed for effects that completely change the initial pattern
-        currentStrip.effect.step( ( arr ) => {
-            currentStrip.arr = arr;
-        } );
-
-        // Set the effect to run continuously
-        currentStrip.effectTimout = setInterval( () => {
-            currentStrip.effect.step( ( arr ) => {
-                currentStrip.arr = arr;
-                markControllersUpdated( stripConfiguration, stripIndex );
-                if ( currentStrip.type === "sharedRender" ) {
-                    currentStrip.applyToStrips( currentStrip.arr );
-                }
-                drawLEDs();
-            } )
-        }, currentStrip.effect.interval );
-    }
-    if ( currentStrip.type === "sharedRender" ) {
-        currentStrip.applyToStrips( currentStrip.arr );
-    }
-}
-
-function setStripDefaults() {
-    currentLEDs.forEach( controller => {
-        controller.forEach( ( strip ) => {
-            if ( strip.default ) {
-                writeConfigToStrips( strip.id, strip.default );
-            } else {
-                blankStrip( strip );
-            }
-        } )
+// LED Code
+function writeConfigToStripsWithContext( stripIndex, options ) {
+    return writeConfigToStrips( stripIndex, options, {
+        currentLEDs,
+        controllersConfig,
+        ledScripts,
+        controllerUpdates,
+        maxFps: MAX_FPS,
+        drawLEDs
     } );
 }
 
-function turnAllLightsOff() {
-    if ( !controllersConfig.length ) {
-        return;
-    }
-    controllersConfig.forEach( ( controller, cIndex ) => {
-            controller.strips.forEach( ( strip, sIndex ) => {
-                controllerUpdates[cIndex] = true;
-                writeConfigToStrips( [ cIndex, sIndex ], {
-                    "trigger": "system",
-                    "pattern": "off",
-                    "patternOptions": {},
-                    "effect": "",
-                    "effectOptions": {}
-                } );
-            } );
-        }
-    );
-    drawLEDs();
-}
-
+// Assign a patter/effect to a strip
 function setLEDs( options ) {
     //clear all other app scripts
     if ( options.trigger === "app" ) {
@@ -661,8 +364,8 @@ function setLEDs( options ) {
             // If sharedRender is set to 'off', set each strip individually
             if ( options.pattern === "off" ) {
                 sharedGroup.strips.forEach( ( stripId ) => {
-                    markControllersUpdated( { controller: stripId[0] }, stripId );
-                    writeConfigToStrips( stripId, options );
+                    markControllersUpdated( { controller: stripId[0] }, stripId, controllerUpdates );
+                    writeConfigToStripsWithContext( stripId, options );
                 } );
                 return;
             }
@@ -679,20 +382,20 @@ function setLEDs( options ) {
             sharedControllers.forEach( controller => {
                 controllerUpdates[controller] = true;
             } );
-            writeConfigToStrips( vStrip, options );
+            writeConfigToStripsWithContext( vStrip, options );
         } else {
             if ( options.trigger === "default" ) {
                 currentLEDs[stripIndex[0]][stripIndex[1]].default = options;
             }
             controllerUpdates[stripIndex[0]] = true;
-            writeConfigToStrips( stripIndex, options );
+            writeConfigToStripsWithContext( stripIndex, options );
         }
     } );
     // After all the strips are set, draw the colors to the strip
     drawLEDs();
 }
 
-//function that handles all writing to the LEDs
+// Final layer to pass LED colors to the controller
 function drawLEDs() {
     if ( drawTimeout ) {
         drawOnTimeout = true;
@@ -733,6 +436,61 @@ function drawLEDs() {
             }
         }
     } );
+}
+
+// LED helper functions
+function clearAppConfigs( noClear ) {
+    noClear = noClear.map( s => stripIndexTextKey( s ) ) ?? [];
+    currentLEDs.forEach( ( controller, cIndex ) => {
+        controller.forEach( ( strip, sIndex ) => {
+            if ( strip.trigger === "app" && !noClear.includes( stripIndexTextKey( [ cIndex, sIndex ] ) ) ) {
+                clearStrip( strip );
+                disableVirtualStripSegments( [ cIndex, sIndex ] );
+                controllerUpdates[cIndex] = true;
+            }
+        } );
+    } );
+}
+
+function setStripDefaults() {
+    currentLEDs.forEach( controller => {
+        controller.forEach( ( strip ) => {
+            if ( strip.default ) {
+                writeConfigToStripsWithContext( strip.id, strip.default );
+            } else {
+                blankStrip( strip );
+            }
+        } )
+    } );
+}
+
+function turnAllLightsOff() {
+    if ( !controllersConfig.length ) {
+        return;
+    }
+    controllersConfig.forEach( ( controller, cIndex ) => {
+            controller.strips.forEach( ( strip, sIndex ) => {
+                controllerUpdates[cIndex] = true;
+                writeConfigToStripsWithContext( [ cIndex, sIndex ], {
+                    "trigger": "system",
+                    "pattern": "off",
+                    "patternOptions": {},
+                    "effect": "",
+                    "effectOptions": {}
+                } );
+            } );
+        }
+    );
+    drawLEDs();
+}
+
+// Matrix
+let matrixInterval;
+
+if ( features.matrixDisplay ) {
+    setTimeout( () => {
+        changeMatrix( { 'id': displayMatrix.default } )
+    }, 500 );
 }
 
 function changeMatrix( options ) {
